Issue #1451 CollectorJsonUtils for 1.0 now also uses GenericData. Also put the common methods between CollectorJsonUtil 1.1 and 1.0 in CollectorJsonHelpers
