+++ dev/com.ibm.ws.logging/src/com/ibm/ws/logging/collector/CollectorJsonHelpers.java
@@ -1,0 +1,207 @@
+/*******************************************************************************
+ * Copyright (c) 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.logging.collector;
+
+import java.text.SimpleDateFormat;
+
+import com.ibm.ws.logging.data.KeyValuePair;
+import com.ibm.ws.logging.data.LogFieldConstants;
+
+/**
+ *
+ */
+public class CollectorJsonHelpers {
+
+    protected static String getEventType(String source, String location) {
+        if (source.equals(CollectorConstants.GC_SOURCE) && location.equals(CollectorConstants.MEMORY)) {
+            return CollectorConstants.GC_EVENT_TYPE;
+        } else if (source.equals(CollectorConstants.MESSAGES_SOURCE) && location.equals(CollectorConstants.MEMORY)) {
+            return CollectorConstants.MESSAGES_LOG_EVENT_TYPE;
+        } else if (source.endsWith(CollectorConstants.TRACE_SOURCE) && location.equals(CollectorConstants.MEMORY)) {
+            return CollectorConstants.TRACE_LOG_EVENT_TYPE;
+        } else if (source.endsWith(CollectorConstants.FFDC_SOURCE) && location.equals(CollectorConstants.MEMORY)) {
+            return CollectorConstants.FFDC_EVENT_TYPE;
+        } else if (source.endsWith(CollectorConstants.ACCESS_LOG_SOURCE) && location.equals(CollectorConstants.MEMORY)) {
+            return CollectorConstants.ACCESS_LOG_EVENT_TYPE;
+        } else
+            return "";
+    }
+
+    protected static ThreadLocal<BurstDateFormat> dateFormatTL = new ThreadLocal<BurstDateFormat>() {
+        @Override
+        protected BurstDateFormat initialValue() {
+            return new BurstDateFormat(new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ"));
+        }
+    };
+
+    protected static boolean addToJSON(StringBuilder sb, String name, String value, boolean jsonEscapeName,
+                                       boolean jsonEscapeValue, boolean trim, boolean isFirstField) {
+
+        boolean b = addToJSON(sb, name, value, jsonEscapeName, jsonEscapeValue, trim, isFirstField, null);
+        return b;
+    }
+
+    protected static boolean addToJSON(StringBuilder sb, String name, String value, boolean jsonEscapeName,
+                                       boolean jsonEscapeValue, boolean trim, boolean isFirstField, KeyValuePair kvp) {
+
+        // if name or value is null just return
+        if (name == null || value == null)
+            return false;
+
+        // add comma if isFirstField == false
+        if (!isFirstField)
+            sb.append(",");
+
+        // trim value if requested
+        if (trim)
+            value = value.trim();
+
+        sb.append("\"");
+        // escape name if requested
+
+        if (jsonEscapeName)
+            jsonEscape3(sb, name);
+        else
+            sb.append(name);
+
+        if (kvp != null && !kvp.getKey().equals("ibm_datetime") && kvp.getType() == KeyValuePair.ValueTypes.NUMBER) {
+
+            sb.append("\":");
+
+            if (jsonEscapeValue)
+                jsonEscape3(sb, value);
+            else
+                sb.append(value);
+
+        } else {
+
+            sb.append("\":\"");
+
+            // escape value if requested
+            if (jsonEscapeValue)
+                jsonEscape3(sb, value);
+            else
+                sb.append(value);
+
+            sb.append("\"");
+
+        }
+        return true;
+    }
+
+    /**
+     * Escape \b, \f, \n, \r, \t, ", \, / characters and appends to a string builder
+     *
+     * @param sb String builder to append to
+     * @param s String to escape
+     */
+    protected static void jsonEscape3(StringBuilder sb, String s) {
+        for (int i = 0; i < s.length(); i++) {
+            char c = s.charAt(i);
+            switch (c) {
+                case '\b':
+                    sb.append("\\b");
+                    break;
+                case '\f':
+                    sb.append("\\f");
+                    break;
+                case '\n':
+                    sb.append("\\n");
+                    break;
+                case '\r':
+                    sb.append("\\r");
+                    break;
+                case '\t':
+                    sb.append("\\t");
+                    break;
+
+                // Fall through because we just need to add \ (escaped) before the character
+                case '\\':
+                case '\"':
+                case '/':
+                    sb.append("\\");
+                    sb.append(c);
+                    break;
+                default:
+                    sb.append(c);
+            }
+        }
+    }
+
+    protected static boolean addCommonFields(StringBuilder sb, String hostName, String wlpUserDir, String serverName,
+                                             boolean isFirstField, String eventType) {
+
+        isFirstField = isFirstField & !addToJSON(sb, "type", eventType, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "hostName", hostName, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "wlpUserDir", wlpUserDir, false, true, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "serverName", serverName, false, false, false, isFirstField);
+
+        return isFirstField;
+    }
+
+    protected static boolean addCommonFieldsGC(StringBuilder sb, String hostName, String wlpUserDir, String serverName, long timestamp, String sequenceNum,
+                                               boolean isFirstField, String eventType) {
+        String datetime = dateFormatTL.get().format(timestamp);
+
+        /* Common fields for all event types */
+
+        isFirstField = isFirstField & !addToJSON(sb, "ibm_datetime", datetime, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "type", eventType, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "host", hostName, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "ibm_userDir", wlpUserDir, false, true, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "ibm_serverName", serverName, false, false, false, isFirstField);
+        isFirstField = isFirstField & !addToJSON(sb, "ibm_sequence", sequenceNum, false, false, false, isFirstField);
+        return isFirstField;
+    }
+
+    protected static String formatMessage(String message, int maxLength) {
+        return (message.length() > maxLength && maxLength > 0) ? message.substring(0, maxLength) + "..." : message;
+    }
+
+    protected static String removeIBMTag(String s) {
+        s = s.replace(LogFieldConstants.IBM_TAG, "");
+        return s;
+    }
+
+    protected static StringBuilder addTagNameForVersion(StringBuilder sb) {
+
+        sb.append(",\"tags\":");
+
+        return sb;
+    }
+
+    protected static String jsonifyTags(String[] tags) {
+        StringBuilder sb = new StringBuilder();
+
+        sb.append("[");
+        for (int i = 0; i < tags.length; i++) {
+
+            tags[i] = tags[i].trim();
+            if (tags[i].contains(" ") || tags[i].contains("-")) {
+                continue;
+            }
+            sb.append("\"");
+            jsonEscape3(sb, tags[i]);
+            sb.append("\"");
+            if (i != tags.length - 1) {
+                sb.append(",");
+            }
+        }
+
+        //Check if have extra comma due to last tag being dropped for
+        if (sb.toString().lastIndexOf(",") == sb.toString().length() - 1) {
+            sb.delete(sb.toString().lastIndexOf(","), sb.toString().lastIndexOf(",") + 1);
+        }
+        sb.append("]");
+        return sb.toString();
+    }
+
+}
