+++ dev/com.ibm.ws.logging/src/com/ibm/ws/logging/internal/impl/JsonLogHandler.java
@@ -20,7 +20,7 @@ import java.util.List;
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.ws.logging.collector.CollectorConstants;
-import com.ibm.ws.logging.collector.CollectorJsonUtils11;
+import com.ibm.ws.logging.collector.CollectorJsonUtils;
 import com.ibm.ws.logging.collector.Formatter;
 import com.ibm.ws.logging.data.GenericData;
 import com.ibm.wsspi.collector.manager.BufferManager;
@@ -158,13 +158,9 @@ public abstract class JsonLogHandler implements SynchronousHandler, Formatter {
     @Override
     public Object formatEvent(String source, String location, Object event, String[] tags, int maxFieldLength) {
 
-//        String eventType = CollectorJsonUtils.getEventType(source, location);
-//        String jsonStr = CollectorJsonUtils.jsonifyEvent(event, eventType, serverName, wlpUserDir, serverHostName, "1.1", tags,
-//                                                         MAXFIELDLENGTH);
-
-        String eventType = CollectorJsonUtils11.getEventType(source, location);
-        String jsonStr = CollectorJsonUtils11.jsonifyEvent(event, eventType, serverName, wlpUserDir, serverHostName, tags,
-                                                           MAXFIELDLENGTH);
+        String eventType = CollectorJsonUtils.getEventType(source, location);
+        String jsonStr = CollectorJsonUtils.jsonifyEvent(event, eventType, serverName, wlpUserDir, serverHostName, "1.0", tags,
+                                                         MAXFIELDLENGTH);
         return jsonStr;
     }
 
@@ -205,24 +201,6 @@ public abstract class JsonLogHandler implements SynchronousHandler, Formatter {
             return CollectorConstants.ACCESS_LOG_SOURCE;
         return "";
     }
-
-    /*
-     * Get the type of source based on the object given by comparing it to the *Data classess
-     */
-//    protected String getSourceTypeFromDataObject(Object event) {
-//        if (event instanceof MessageLogData) {
-//            return CollectorConstants.MESSAGES_SOURCE;
-//        } else if (event instanceof TraceLogData) {
-//            return CollectorConstants.TRACE_SOURCE;
-//        } else if (event instanceof AccessLogData) {
-//            return CollectorConstants.ACCESS_LOG_SOURCE;
-//        } else if (event instanceof FFDCData) {
-//            return CollectorConstants.FFDC_SOURCE;
-//        } else {
-//            return "";
-//        }
-//
-//    }
 
     protected String getSourceTypeFromDataObject(Object event) {
